+++ CHANGELOG.md
@@ -5,12 +5,15 @@
 
 ### Changed
 ### Removed
+- Dropped support for Python 3.9
 ### Added
+- Added support for parsing finer details in multi-objective logs (#GH36)
 - Added the `warnings_action` parameter to `gurobi_logtools.parse` (#GH88)
 - NumConVars, NumIntVars and NumBinVars are added to capture variable types in original model (#GH89)
 - Added parsing for parameters under "Non-default parameters" section in logs (#GH92)
 - Added parsing for PDHG (#GH60)
 - Added detailed parsing for crossover steps (#GH106)
+- Added support for Python 3.14
 
 ## 4.0.1 - 2025-08-14
 ### Fixed